feat(infra): encrypt lambda environment variables with kms cmk
Each lambda construct resolves FlexEncryptionKey from SSM and applies
environmentEncryption inline. Standalone Construct classes (no inheritance)
preserve original construct paths, avoiding CloudFormation logical ID
changes that caused naming conflicts.

- FlexPublicFunction, FlexPrivateIsolatedFunction: alarm ID Alarm
- FlexPrivateEgressFunction: alarm ID Alarm (unchanged)
- Shared resolveEncryptionKey() utility in utils/lambda.ts
- Deleted FlexBaseFunction (inheritance-based approach)
- CKV_AWS_173 remains globally suppressed for CDK-internal Lambdas
