+++ platform/infra/flex/src/constructs/lambda/flex-private-egress-function.ts
@@ -1,24 +1,70 @@
+import { getEnvConfig } from "@flex/utils";
+import { Tags } from "aws-cdk-lib";
 import { ISecurityGroup, IVpc, SubnetType } from "aws-cdk-lib/aws-ec2";
+import { Runtime, Tracing } from "aws-cdk-lib/aws-lambda";
+import { NodejsFunction } from "aws-cdk-lib/aws-lambda-nodejs";
+import { LogGroup, RetentionDays } from "aws-cdk-lib/aws-logs";
 import { Construct } from "constructs";
 
-import type { FlexFunctionProps } from "../types";
-import { FlexBaseFunction } from "./flex-base-function";
+import { resolveEncryptionKey } from "../../utils/lambda";
+import { LambdaAlarms } from "../alarms/lambda";
+import { FlexFunctionProps } from "../types";
+
+const { stage } = getEnvConfig();
 
 interface FlexPrivateEgressFunctionProps extends FlexFunctionProps {
   vpc: IVpc;
   privateEgressSg: ISecurityGroup;
 }
 
-export class FlexPrivateEgressFunction extends FlexBaseFunction {
+export class FlexPrivateEgressFunction extends Construct {
+  public readonly function: NodejsFunction;
+
   constructor(
     scope: Construct,
     id: string,
-    { vpc, privateEgressSg, ...props }: FlexPrivateEgressFunctionProps,
-  ) {
-    super(scope, id, props, {
+    {
       vpc,
-      securityGroups: [privateEgressSg],
-      vpcSubnets: { subnetType: SubnetType.PRIVATE_WITH_EGRESS },
+      privateEgressSg,
+      criticalAction,
+      warningAction,
+      ...functionProps
+    }: FlexPrivateEgressFunctionProps,
+  ) {
+    super(scope, id);
+
+    const logGroup = new LogGroup(this, "LogGroup", {
+      retention: RetentionDays.ONE_YEAR,
     });
+
+    this.function = new NodejsFunction(this, "Function", {
+      runtime: Runtime.NODEJS_24_X,
+      tracing: Tracing.ACTIVE,
+      ...functionProps,
+      environmentEncryption: resolveEncryptionKey(this),
+      environment: {
+        ...functionProps.environment,
+        FLEX_ENVIRONMENT: stage,
+      },
+      logGroup,
+      securityGroups: [privateEgressSg],
+      vpc,
+      vpcSubnets: {
+        subnetType: SubnetType.PRIVATE_WITH_EGRESS,
+      },
+    });
+
+    new LambdaAlarms(this, "Alarm", {
+      fn: this.function,
+      alarmNamePrefix: `${stage}-${id.toLowerCase()}-alarm`,
+      criticalAction,
+      warningAction,
+    });
+
+    if (functionProps.domain) {
+      Tags.of(this.function).add("ResourceOwner", functionProps.domain, {
+        priority: 200,
+      });
+    }
   }
 }
